GH-35

+++ application/models/DAO/PeriodDAO_model.php
@@ -3,7 +3,8 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class PeriodDAO_model extends CI_Model {
 
-	var $periodTable = 'Period'; // Table name 
+    var $periodTable = 'Period';
+    var $formTable = 'Form';
 
 	function __construct()
 	{
@@ -35,12 +36,47 @@ class PeriodDAO_model extends CI_Model {
     }
 
     /****************************************
+    - Check if there is a unique period with number and year
+    ****************************************/
+    private function validatePeriod($pNumber, $pYear)
+    {
+        $this->db->from($this->periodTable);
+        $array = array('number' => $pNumber, 'year' => $pYear);
+        $this->db->where($array);
+        $query = $this->db->get();
+        return $query->num_rows();
+    }
+
+    /****************************************
+    - Check if there is a unique period associated with a form
+    ****************************************/
+    private function validatePeriodInForms($pId)
+    {
+        $this->db->from($this->formTable);
+        $this->db->where('idPeriod', $pId);
+        $query = $this->db->get();
+        return $query->num_rows();
+    }
+
+    /****************************************
     - Insert the new period in the database.
     ****************************************/
     public function insert($pPeriod)
     {
-        $this->db->insert($this->periodTable, $pPeriod);
-        return $this->db->insert_id();
+        $validate = $this->validatePeriod($pPeriod['number'], $pPeriod['year']);
+
+        if($validate == 0)
+        {
+            $this->db->insert($this->periodTable, $pPeriod);
+            echo 'true';
+            return;
+        }
+
+        else
+        {
+            echo 'false';
+            return;
+        }
     }
 
     /****************************************
@@ -48,13 +84,26 @@ class PeriodDAO_model extends CI_Model {
     ****************************************/
     public function edit($pPeriod)
     {
-        $changes = array(
-            'number' => $pPeriod['number'],
-            'year' => $pPeriod['year']
-        );
-        $this->db->where('idPeriod', $pPeriod['idPeriod']);
-        $this->db->update($this->periodTable, $changes);
-        return $this->db->affected_rows();
+        $validatePeriod = $this->validatePeriod($pPeriod['number'], $pPeriod['year']);
+        $validatePeriodInForms = $this->validatePeriodInForms($pPeriod['idPeriod']);
+
+        if($validatePeriod == 0 && $validatePeriodInForms == 0)
+        {
+            $changes = array(
+                'number' => $pPeriod['number'],
+                'year' => $pPeriod['year']
+            );
+
+            $this->db->where('idPeriod', $pPeriod['idPeriod']);
+            $this->db->update($this->periodTable, $changes);
+            echo 'true';
+            return;
+        }
+        else
+        {
+            echo 'false';
+            return;
+        }
     }
 
 
@@ -63,8 +112,21 @@ class PeriodDAO_model extends CI_Model {
     ****************************************/
     public function delete($pId)
     {
-        $this->db->where('idPeriod', $pId);
-        $this->db->delete($this->periodTable);
+        $validatePeriodInForms = $this->validatePeriodInForms($pId);
+
+        if($validatePeriodInForms == 0)
+        {
+            $this->db->where('idPeriod', $pId);
+            $this->db->delete($this->periodTable);
+            echo 'true';
+            return;
+        }
+
+        else
+        {
+            echo 'false';
+            return;
+        }
     }
 
 }
